backend/requirements.txt
asgiref==3.5.2
astroid==2.12.12
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
boto3==1.26.10
botocore==1.29.10
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
coveralls==3.3.1
cryptography==38.0.3
daphne==4.0.0
Deprecated==1.2.13
dill==0.3.6
dj-rest-auth==2.2.5
Django==4.1.2
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.4
mysqlclient==2.1.1
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.2
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pylint==2.15.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.5
Twisted==22.10.0
txaio==22.2.1
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
zope.interface==5.5.1
